requirements.txt
# --- Core Scraping & Crawling ---
requests==2.32.3
beautifulsoup4==4.13.3
lxml==5.3.0

# --- RAG & Vector Database ---
chromadb>=0.5.20
python-dotenv==1.0.1
rank_bm25==0.2.2

# --- Backend & UI ---
fastapi==0.115.6
uvicorn==0.34.0
streamlit==1.52.0
openai>=1.0.0